+++ src/main/java/com/sanitas/calculator/response/CalculatorResponse.java
@@ -1,5 +1,7 @@
 package com.sanitas.calculator.response;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -7,6 +9,6 @@ public class CalculatorResponse {
 	
 	@Getter
 	@Setter
-	private Integer result;
+	private BigDecimal result;
 
 }
